% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowDM.r
\name{lowDM}
\alias{lowDM}
\title{Title}
\usage{
lowDM(
  Glmat,
  method = "phred",
  userFun,
  adjust = TRUE,
  adjust.param = c(25, 50)
)
}
\arguments{
\item{Glmat}{}

\item{method}{}

\item{userFun}{}

\item{adjust}{}

\item{adjust.param}{}
}
\value{

}
\description{
Title
}
