% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowKi.r
\name{lowKi}
\alias{lowKi}
\title{Computing Kinship and Fraternity matrices}
\usage{
lowKi(
  filename,
  field = c("PL", "GP"),
  adjust = TRUE,
  fraternity = FALSE,
  adjust.par = c(20, 20, 10)
)
}
\arguments{
\item{filename}{path to a VCF file}

\item{field}{which field of the VCF file to use}

\item{adjust}{logical. TRUE to use adjustment method}

\item{fraternity}{logical. TRUE to compute the fraternity matrix instead of the kinship matrix}

\item{adjust.par}{Parameters for the adjustment.}
}
\value{
A symmetric matrix.
}
\description{
Computing Kinship and Fraternity matrices
}
\details{
This function computes the kinship matrix or the fraternity matrix (aka as 'dominance matrix')
from shallow sequencing data. It relies on moment estimates and on an adjustment method which allows
to correct the bias.

The method is exposed in details in a yet unpublished paper. The three values in `adjust.par`
allow to control the number of individuals used in the adjustment. The first (resp. second) value gives the number 
of pairs of individuals with the lowest (resp. highest) kinship to be selected; the third gives the number of
random individuals to be added.

Note that all SNPs present in the VCF file will be used, regardless of the MAF or of the chromosome number.
}
