% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowKin.r
\name{lowKin}
\alias{lowKin}
\title{Title}
\usage{
lowKin(
  Glmat,
  method = c("none", "imputation", "phred"),
  adjust = TRUE,
  dominance = FALSE,
  constraint = FALSE
)
}
\arguments{
\item{Glmat}{}

\item{method}{}

\item{adjust}{}

\item{constraint}{constraint regression parameters of v1 and v2 to be equal}
}
\value{

}
\description{
Title
}
