% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRM.vcf.r
\name{lowKin.vcf}
\alias{lowKin.vcf}
\title{Title}
\usage{
lowKin.vcf(
  filename,
  field = c("PL", "GP"),
  adjust = TRUE,
  dominance = FALSE,
  constraint = TRUE
)
}
\arguments{
\item{filename}{path to a VCF file}

\item{field}{which field of the VCF file to use}

\item{constraint}{constraint regression parameters of v1 and v2 to be equal}
}
\value{

}
\description{
Title
}
\details{
I left the 'constraint' option for the moment but this should disappear
}
