% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowKi.r
\name{lowKi}
\alias{lowKi}
\title{Computing Kinship and Fraternity matrices}
\usage{
lowKi(
  filename,
  field = c("PL", "GP"),
  freqs,
  adjust = TRUE,
  fraternity = FALSE,
  adjust.par = c(20, 20, 10)
)
}
\arguments{
\item{filename}{path to a VCF file}

\item{field}{which field of the VCF file to use}

\item{freqs}{[optional] allele frequencies (see details)}

\item{adjust}{logical. TRUE to use adjustment method}

\item{fraternity}{logical. TRUE to compute the fraternity matrix instead of the kinship matrix}

\item{adjust.par}{Parameters for the adjustment procedure (details below).}
}
\value{
A symmetric matrix.
}
\description{
Computing Kinship and Fraternity matrices
}
\details{
This function computes the kinship matrix or the fraternity matrix (aka as 'dominance matrix')
from shallow sequencing data. It relies on moment estimates and on an adjustment method which allows
to correct the bias.

The method is exposed in details in a yet unpublished paper. The three values in `adjust.par`
allow to control the number of individuals used in the adjustment. The first (resp. second) value gives the number 
of pairs of individuals with the lowest (resp. highest) kinship to be selected; the third gives the number of
random individuals to be added. The adjustment procedure is based on all the estimated coefficients between 
the selected individuals.

All SNPs present in the VCF file will be used, regardless of the MAF or of the chromosome number.
}
\examples{
# see vignette for more examples
vcf.file <- system.file("extdata", "shallow.vcf.gz", package="LowKi")
kinship.file <-system.file("extdata", "kinship.rds", package="LowKi")
K.low.1 <- lowKi(vcf.file)
MLfreq <- vcf.allele.freq(vcf.file)
K.low.2 <- lowKi(vcf.file, freqs = MLfreq)
K.real <- readRDS(kinship.file)
par(mfrow=c(1,2))
plot(K.real, K.low.1, xlab = "true values", ylab = "LowKi estimates")
abline(0,1,col="red")
plot(K.real, K.low.2, xlab = "true values", ylab = "LowKi estimates")
abline(0,1,col="red")
}
