% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowGRMcpp.r
\name{lowKin}
\alias{lowKin}
\title{Title}
\usage{
lowKin(
  Glmat,
  method = c("imputation", "phred"),
  adjust = TRUE,
  dominance = FALSE
)
}
\arguments{
\item{Glmat}{}

\item{method}{}

\item{adjust}{}
}
\value{

}
\description{
Title
}
