% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowKin.vcf.r
\name{lowKin.vcf}
\alias{lowKin.vcf}
\title{Title}
\usage{
lowKin.vcf(
  filename,
  field = c("PL", "GP"),
  adjust = TRUE,
  fraternity = FALSE,
  adjust.par = c(20, 20, 10)
)
}
\arguments{
\item{filename}{path to a VCF file}

\item{field}{which field of the VCF file to use}

\item{fraternity}{}
}
\value{

}
\description{
Title
}
\details{
I left the 'constraint' option for the moment but this should disappear
}
